- Tests readability simplification - Dependencies normalization
