+++ composer.json
@@ -21,14 +21,13 @@
         "symfony/property-access": "~2.3|~3.0",
         "jms/serializer-bundle": "~0.13|~1.0|~2.0",
         "pagerfanta/pagerfanta": "~1.0",
-        "willdurand/hateoas": ">2.0"
+        "willdurand/hateoas": "~2.0",
+        "symfony/expression-language": "~2.4",
+        "symfony/property-access": "^3.3",
+        "doctrine/common": "^2.4.8"
     },
     "require-dev": {
-        "doctrine/orm": "^2.4.8",
-        "symfony/property-access": "^3.3",
-        "symfony/expression-language": "~2.4",
-        "twig/twig": "~1.12",
-        "phpunit/phpunit": "~4.5"
+        "phpunit/phpunit": "~5.7"
     },
     "autoload": {
         "psr-0": {
